Move db datasource to db package
