+++ app/src/main/kotlin/com/ayatk/biblio/data/db/LibraryLocalDataSource.kt
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.ayatk.biblio.data.datasource.library
+package com.ayatk.biblio.data.db
 
 import com.ayatk.biblio.domain.repository.LibraryRepository
 import com.ayatk.biblio.model.Library
